Feature/KAN-185: Auction realtime bidding implementation into develop

See merge request courses/2024-summer/csci-5308/group06!18

+++ Auction_backend/src/main/java/com/online/auction/service/AccountService.java
@@ -3,6 +3,6 @@ package com.online.auction.service;
 import com.online.auction.exception.ServiceException;
 
 public interface AccountService {
-    float getAccountBalance(Integer userId) throws ServiceException;
+    double getAccountBalance(Integer userId) throws ServiceException;
     void addFunds(Integer userId, float amount) throws ServiceException;
 }
